% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nyc_terrain_path}
\alias{nyc_terrain_path}
\title{Get the file path to a bundled NYC terrain GeoTIFF}
\usage{
nyc_terrain_path(resolution = c("100ft", "50ft"))
}
\arguments{
\item{resolution}{Either \code{"100ft"} (default) or \code{"50ft"}.}
}
\value{
A character string file path.
}
\description{
Returns the path to a GeoTIFF bundled with the package. Useful when
you want to read the raster with another tool (e.g. \code{stars}, GDAL,
\code{rayshader}) instead of \code{terra}.
}
\examples{
nyc_terrain_path()
nyc_terrain_path("50ft")

}
\keyword{functions}
